requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
pymongo==4.11.2
python-dotenv==1.0.1
smmap==5.0.2
Werkzeug==3.1.3
WTForms==3.2.1
SQLAlchemy==2.0.36